requirements.txt
django>=3.0
django-subadmin>=2.0
